40-IoT-event-routing/Mike/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
azure.iot.device==3.0.0b2